!F (Schematyc) Introducing Schematyc Beta - an experimental tool to build gameplay systems giving designers the power to create new and reusable functionality from a set of building blocks provided by programmers. Whereas Flowgraphs are great for level scripting, Schematyc is designed to provide more finite control of objects within those levels.  All logic is driven by state and context in order to simplify the information that is presented to designers, and greatly reduced latency should make it possible to take new gameplay systems beyond the prototyping stage without the need to re-write them in C++. (Approved by achim) !XI Copying //ce/dev_schematyc to main (//ce/main) @ CL 1457568 Review: ID 61936 by Author: Paul Slinger, Author: Dominique Mader, Author: Filip Lundgren, Reviewer: Scott Peter, Reviewer: Nico Moss, Observer: Marco Hopp, Observer: Rok Erjavec, Observer: Matthijs, Observer: Samuel Kahn
Copied from Perforce
 Change: 1457647

+++ Code/CryEngine/CryAction/ActionGame.cpp
@@ -3970,7 +3970,7 @@ void CActionGame::RegisterBrokenMesh(IPhysicalEntity* pPhysEnt, IGeometry* pPhys
 					m_totBreakageSize -= iter->second.size;
 					m_mapBrokenMeshes.erase(iter);
 				}
-			uint32 eraseTo = min((size_t)(id + 1), m_brokenMeshRemovals.size());
+			size_t eraseTo = min((size_t)(id + 1), m_brokenMeshRemovals.size());
 			m_brokenMeshRemovals.erase(m_brokenMeshRemovals.begin(), m_brokenMeshRemovals.begin() + eraseTo);
 			return;
 		}
